add style & fix bugs

+++ View/Categories/delete.php
@@ -14,14 +14,17 @@ if ($_SERVER["REQUEST_METHOD"] == "GET") {
     if (empty($_GET["id"])) {
         $errors['id'] = "Category Not Found";
     }
+}
 
+if ($_SERVER["REQUEST_METHOD"] == "POST") {
     // If there are no validation errors, proceed with updating the product
     if (empty($errors)) {
         if ($categoryController->destroy($id)) {
             header("Location: index.php");
-        } else {
-            echo "Error";
+            exit();
         }
+    } else {
+        echo "Error";
     }
 }
 
@@ -56,30 +59,41 @@ if ($_SERVER["REQUEST_METHOD"] == "GET") {
     </style>
 </head>
 
-<body class="container mt-4">
-    <h2>Delete Category</h2>
-    <a href="index.php">Back to Category List</a>
-    <br><br>
-    <?php if (count($row) > 0) : ?>
-        <p>Are you sure you want to delete the following category?</p>
-        <table>
-            <tr>
-                <td>ID:</td>
-                <td><?php echo $row[0]["id"]; ?></td>
-            </tr>
-            <tr>
-                <td>Category Name:</td>
-                <td><?php echo $row[0]["category_name"]; ?></td>
-            </tr>
-            <tr>
-        </table>
-        <form action="" method="get">
-            <input type="hidden" name="id" value="<?php echo $row[0]['id']; ?>">
-            <input type="submit" value="Delete Category">
-        </form>
-    <?php else : ?>
-        <p>Data not found!</p>
-    <?php endif ?>
+<body>
+    <main class="container mt-5" style="max-width: 800px; margin: auto;">
+        <h2 class="mb-5">Delete Category</h2>
+        <a href="index.php" class="btn btn-outline-secondary mb-3">
+            < Back to Category List</a>
+                <br><br>
+                <?php if (count($row) > 0) : ?>
+                    <div class="card">
+                        <div class="card-body">
+                            <h4 class="card-title">Are you sure you want to delete the following category?</h4>
+                            <table class="mb-3">
+                                <tr>
+                                    <td>ID</td>
+                                    <td>: <?php echo $row[0]["id"]; ?></td>
+                                </tr>
+                                <tr>
+                                    <td>Category Name</td>
+                                    <td>: <?php echo $row[0]["category_name"]; ?></td>
+                                </tr>
+                                <tr>
+                            </table>
+                            <form action="" method="post">
+                                <input type="hidden" name="id" value="<?php echo $row[0]['id']; ?>">
+                                <?php if (isset($errors['id'])) : ?>
+                                    <div class="alert alert-danger mt-2" role="alert">
+                                        <?php echo $errors['id']; ?>
+                                    </div>
+                                <?php endif ?>
+                                <input type="submit" value="Delete Category" class="btn btn-danger w-100">
+                            </form>
+                        </div>
+                    </div>
+                <?php else : ?>
+                    <p class="alert alert-danger mt-2">Data not found!</p>
+                <?php endif ?>
 </body>
 
 </html>
